Created pr create subcommand (reused pull-request command code) Refactored pr subcommands
