+++ commands/pr.go
@@ -14,6 +14,7 @@ var (
 	cmdPr = &Command{
 		Run: printHelp,
 		Usage: `
+pr create [-focp] [-b <BASE>] [-h <HEAD>] [-r <REVIEWERS> ] [-a <ASSIGNEES>] [-M <MILESTONE>] [-l <LABELS>]
 pr list [-s <STATE>] [-h <HEAD>] [-b <BASE>] [-o <SORT_KEY> [-^]] [-f <FORMAT>] [-L <LIMIT>]
 pr checkout <PR-NUMBER> [<BRANCH>]
 pr info <PR-NUMBER>
@@ -30,6 +31,9 @@ pr info <PR-NUMBER>
 
 	* _info_:
 		Show info about given pull request.
+
+	* _create_:
+		Create a pull request
 
 ## Options:
 
@@ -158,12 +162,24 @@ hub-issue(1), hub-pull-request(1), hub(1)
 		Key: "info",
 		Run: prInfo,
 	}
+
+	cmdCreatePr = &Command{
+		Key: "create",
+		Run: createPr,
+		Usage: `
+pr create [-focp] [-b <BASE>] [-h <HEAD>] [-r <REVIEWERS> ] [-a <ASSIGNEES>] [-M <MILESTONE>] [-l <LABELS>]                                                                                                   
+pr create -m <MESSAGE> [--edit]
+pr create -F <FILE> [--edit]
+pr create -i <ISSUE>
+`,
+	}
 )
 
 func init() {
 	cmdPr.Use(cmdListPulls)
 	cmdPr.Use(cmdCheckoutPr)
 	cmdPr.Use(cmdPrInfo)
+	cmdPr.Use(cmdCreatePr)
 	CmdRunner.Use(cmdPr)
 }
 
@@ -262,48 +278,6 @@ func checkoutPr(command *Command, args *Args) {
 	utils.Check(err)
 
 	args.Replace(args.Executable, "checkout", newArgs...)
-}
-
-func prInfo(command *Command, args *Args) {
-	localRepo, err := github.LocalRepo()
-	utils.Check(err)
-
-	project, err := localRepo.MainProject()
-	utils.Check(err)
-
-	gh := github.NewClient(project.Host)
-
-	args.NoForward()
-	words := args.Words()
-	if len(words) == 0 {
-		utils.Check(fmt.Errorf("Error: no pull request number given"))
-	}
-
-	prNumberString := words[0]
-	pr, err := gh.PullRequest(project, prNumberString)
-	utils.Check(err)
-	flagPullRequestFormat := args.Flag.Value("--format")
-	if !args.Flag.HasReceived("--format") {
-		flagPullRequestFormat = "%pC%>(8)%i%Creset  %t%  l%nRequested reviewers: %rs%n%n%b%n"
-	}
-	colorize := colorizeOutput(args.Flag.HasReceived("--color"), args.Flag.Value("--color"))
-	ui.Print(formatPullRequest(*pr, flagPullRequestFormat, colorize))
-	comments, err := gh.FetchPRComments(project, prNumberString)
-	utils.Check(err)
-	reviewsToComments := make(map[int]int)
-	for i, comment := range comments {
-		reviewsToComments[comment.ReviewId] = i
-	}
-	reviews, err := gh.FetchReviews(project, prNumberString)
-	utils.Check(err)
-	for _, review := range reviews {
-		ui.Printf("%s %s on %s\n", review.User.Login, review.State, review.CreatedAt)
-		if review.State != "APPROVED" {
-			comment := comments[reviewsToComments[review.Id]]
-			ui.Printf("%s\n\n%s\n", comment.DiffHunk, comment.Body)
-		}
-		ui.Printf("\n")
-	}
 }
 
 func formatPullRequest(pr github.PullRequest, format string, colorize bool) string {
